add bruteforce day 5 pt. 2
