+++ Cargo.toml
@@ -27,6 +27,10 @@ path = "src/bin/day4/day4.rs"
 name = "day5-1"
 path = "src/bin/day5/day5-1.rs"
 
+[[bin]]
+name = "day5-2"
+path = "src/bin/day5/day5-2.rs"
+
 [dependencies]
 itertools = "0.12.0"
 num_cpus = "1.16.0"
